Line of Sight Updates
Used buffering on the endpoints of leg1 and leg2 of the triangle connecting two surfaces to fix a case that wasn't allowing the triangle legs to be updated (messy, ugly fix, but works for now). Also changed the "elif" statements for legs 1 and 2 to "ifs" (code wasn't running correctly. As far as I can see, LOS works as it should for the shaded geometry.