More LoS fixes - Midpoint and Circles
LoS fix for checking buffer area around the midpoint for incorrect intersection/adjustment, as the incorrect intersection with the outside may not always be exactly at the midpoint.

Additional fix for the distribution circles being too large and jutting out of the geometry (needs further testing to make sure it works for all cases though, but for now it has fixed the most glaringly obvious issues).